Correctly generate code for the avro ops.
We remove avro_helper.h and avro_helper.cc, and put their contents
into parse_avro_record.cc because the tf_gen_op_libs is not allowed to
declare dependencies on other .h and .cc files, for whatever odd
reason.
